+++ src/components/cookifyInput.tsx
@@ -3,11 +3,11 @@ import { useCookifyProvider } from '../context/cookifyContext'
 import { CookifyInputProps } from '../types'
 
 export const CookifyInput: React.FC<CookifyInputProps> = (props) => {
-    const {consentObject, actionCheckbox} = useCookifyProvider();
+    const {consentObject, actionCheckbox} = useCookifyProvider()
 
     const handleChange: React.ChangeEventHandler<HTMLInputElement> = (e) => {
         actionCheckbox(e.currentTarget.name)
-    };
+    }
 
     return (
         <input
@@ -15,7 +15,7 @@ export const CookifyInput: React.FC<CookifyInputProps> = (props) => {
             checked={consentObject.data[props.name]}
             onChange={e => handleChange(e)}
         />
-    );
-};
+    )
+}
 
 export default CookifyInput